src/requirements.txt
functions-framework
google-cloud-storage
google-cloud-vision
google-cloud-firestore
flask